moved SlotAndBlockRoot to :ethereum:datastructures to remove the dependency on :storage:api
