+++ ethereum/datastructures/src/main/java/tech/pegasys/teku/datastructures/forkchoice/MutableStore.java
@@ -17,9 +17,9 @@ import com.google.common.primitives.UnsignedLong;
 import org.apache.tuweni.bytes.Bytes32;
 import tech.pegasys.teku.datastructures.blocks.SignedBeaconBlock;
 import tech.pegasys.teku.datastructures.blocks.SignedBlockAndState;
+import tech.pegasys.teku.datastructures.blocks.SlotAndBlockRoot;
 import tech.pegasys.teku.datastructures.state.BeaconState;
 import tech.pegasys.teku.datastructures.state.Checkpoint;
-import tech.pegasys.teku.storage.api.schema.SlotAndBlockRoot;
 
 public interface MutableStore extends ReadOnlyStore {
